requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchinfo
tqdm
pillow
pandas
matplotlib
jupyter
albumentations
scikit-learn
SimpleITK
rich